# apps/web/openapi.json
{"openapi":"3.1.0","info":{"title":"myboard API","version":"0.1.0"},"paths":{"/api/health":{"get":{"tags":["health"],"summary":"Health","description":"Процесс отвечает и база данных доступна.","operationId":"health_api_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Health"}}}}}}},"/api/admin/login":{"post":{"tags":["admin"],"summary":"Login","description":"ADM-01: вход в панель отдельной учётной записью администратора.","operationId":"login_api_admin_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Credentials"}}},"required":true},"responses":{"204":{"description":"Successful Response"},"401":{"description":"Invalid email or password"},"429":{"description":"Too many sign-in attempts, try again later"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/admin/logout":{"post":{"tags":["admin"],"summary":"Logout","operationId":"logout_api_admin_logout_post","responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/admin/session":{"get":{"tags":["admin"],"summary":"Get Session","description":"Вошёл ли браузер в панель. Отвечает 200 и без сессии — интерфейс решает, что показать.","operationId":"get_session_api_admin_session_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdminSession"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/admin/users":{"get":{"tags":["admin"],"summary":"List Users","description":"ADM-02: список учётных записей пользователей досок.","operationId":"list_users_api_admin_users_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/UserOut"},"type":"array","title":"Response List Users Api Admin Users Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["admin"],"summary":"Create User","description":"ADM-03: новая учётка с именем, почтой и паролем; ADM-07: почта не повторяется.","operationId":"create_user_api_admin_users_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserOut"}}}},"409":{"description":"Email is already in use"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/admin/users/{user_id}":{"patch":{"tags":["admin"],"summary":"Update User","description":"ADM-04: имя, почта, пароль; ADM-05/ADM-06: `disabled` отключает и включает учётку.","operationId":"update_user_api_admin_users__user_id__patch","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"User Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserOut"}}}},"404":{"description":"User not found"},"409":{"description":"Email is already in use"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/login":{"post":{"tags":["account"],"summary":"Login","description":"ACC-01: вход по почте и паролю, заданным администратором.\n\nACC-02: неизвестная почта, неверный пароль и отключённая учётка дают один и тот же 401.","operationId":"login_api_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Credentials"}}},"required":true},"responses":{"204":{"description":"Successful Response"},"401":{"description":"Invalid email or password"},"429":{"description":"Too many sign-in attempts, try again later"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/logout":{"post":{"tags":["account"],"summary":"Logout","description":"ACC-03: выход отзывает сессию; скопированная cookie больше не действует.","operationId":"logout_api_logout_post","responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/session":{"get":{"tags":["account"],"summary":"Get Session","description":"Вошёл ли браузер. Отвечает 200 и без сессии; живую сессию продлевает (ACC-03).","operationId":"get_session_api_session_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountSession"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/boards":{"get":{"tags":["library"],"summary":"List Boards","description":"ACC-04, BRD-04: полный список досок пользователя; BRD-05, BRD-06: порядок, фильтр, поиск.","operationId":"list_boards_api_boards_get","parameters":[{"name":"q","in":"query","required":false,"schema":{"anyOf":[{"type":"string","maxLength":200},{"type":"null"}],"description":"Часть названия","title":"Q"},"description":"Часть названия"},{"name":"sort","in":"query","required":false,"schema":{"$ref":"#/components/schemas/BoardSort","default":"updated"}},{"name":"modified_since","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"description":"Только доски, изменённые с этого момента","title":"Modified Since"},"description":"Только доски, изменённые с этого момента"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/BoardOut"},"title":"Response List Boards Api Boards Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["library"],"summary":"Create Board","description":"BRD-01: новая доска сразу попадает в список; интерфейс открывает её по `id`.","operationId":"create_board_api_boards_post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BoardCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BoardOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/boards/recent":{"get":{"tags":["library"],"summary":"Recent Boards","description":"BRD-04: недавние доски — последние изменённые, новые сверху.","operationId":"recent_boards_api_boards_recent_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/BoardOut"},"type":"array","title":"Response Recent Boards Api Boards Recent Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/boards/{board_id}":{"get":{"tags":["library"],"summary":"Get Board","operationId":"get_board_api_boards__board_id__get","parameters":[{"name":"board_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Board Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BoardOut"}}}},"404":{"description":"Board not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"tags":["library"],"summary":"Rename Board","description":"BRD-02","operationId":"rename_board_api_boards__board_id__patch","parameters":[{"name":"board_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Board Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BoardRename"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BoardOut"}}}},"404":{"description":"Board not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["library"],"summary":"Delete Board","description":"BRD-03","operationId":"delete_board_api_boards__board_id__delete","parameters":[{"name":"board_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Board Id"}}],"responses":{"204":{"description":"Successful Response"},"404":{"description":"Board not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/boards/{board_id}/folder":{"put":{"tags":["library"],"summary":"Move Board","description":"BRD-10: перенести доску в папку (`folder_id`) или на верхний уровень (`null`).","operationId":"move_board_api_boards__board_id__folder_put","parameters":[{"name":"board_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Board Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BoardMove"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BoardOut"}}}},"404":{"description":"Board or folder not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/boards/{board_id}/favorite":{"put":{"tags":["library"],"summary":"Add Favorite","description":"BRD-07: добавить доску в избранное.","operationId":"add_favorite_api_boards__board_id__favorite_put","parameters":[{"name":"board_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Board Id"}}],"responses":{"204":{"description":"Successful Response"},"404":{"description":"Board not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["library"],"summary":"Remove Favorite","description":"BRD-07: убрать доску из избранного.","operationId":"remove_favorite_api_boards__board_id__favorite_delete","parameters":[{"name":"board_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Board Id"}}],"responses":{"204":{"description":"Successful Response"},"404":{"description":"Board not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/folders":{"get":{"tags":["library"],"summary":"List Folders","description":"BRD-09: все папки пользователя (дерево строит интерфейс); BRD-06: поиск по названию.","operationId":"list_folders_api_folders_get","parameters":[{"name":"q","in":"query","required":false,"schema":{"anyOf":[{"type":"string","maxLength":200},{"type":"null"}],"description":"Часть названия","title":"Q"},"description":"Часть названия"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/FolderOut"},"title":"Response List Folders Api Folders Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["library"],"summary":"Create Folder","description":"BRD-09: папка на верхнем уровне или внутри `parent_id`.","operationId":"create_folder_api_folders_post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FolderCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FolderOut"}}}},"404":{"description":"Folder not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/folders/{folder_id}/position":{"put":{"tags":["library"],"summary":"Move Folder","description":"BRD-10: поставить папку на место `position` среди папок `parent_id` (null — верхний\nуровень). Вложить папку в неё саму или в её дочернюю нельзя — 409.","operationId":"move_folder_api_folders__folder_id__position_put","parameters":[{"name":"folder_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Folder Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FolderMove"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FolderOut"}}}},"404":{"description":"Folder not found"},"409":{"description":"A folder cannot be moved into itself or its subfolder"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/folders/{folder_id}/favorite":{"put":{"tags":["library"],"summary":"Add Favorite","description":"BRD-07: добавить папку в избранное.","operationId":"add_favorite_api_folders__folder_id__favorite_put","parameters":[{"name":"folder_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Folder Id"}}],"responses":{"204":{"description":"Successful Response"},"404":{"description":"Folder not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["library"],"summary":"Remove Favorite","description":"BRD-07: убрать папку из избранного.","operationId":"remove_favorite_api_folders__folder_id__favorite_delete","parameters":[{"name":"folder_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Folder Id"}}],"responses":{"204":{"description":"Successful Response"},"404":{"description":"Folder not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"AccountSession":{"properties":{"authenticated":{"type":"boolean","title":"Authenticated"},"name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"},"email":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Email"}},"type":"object","required":["authenticated","name","email"],"title":"AccountSession","description":"Состояние входа пользователя досок: имя и почта или `null` (ACC-01, ACC-03)."},"AdminSession":{"properties":{"authenticated":{"type":"boolean","title":"Authenticated"},"email":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Email"}},"type":"object","required":["authenticated","email"],"title":"AdminSession","description":"Состояние входа в панель: `email` администратора или `null`."},"BoardCreate":{"properties":{"title":{"anyOf":[{"type":"string","maxLength":200,"minLength":1},{"type":"null"}],"title":"Title"}},"additionalProperties":false,"type":"object","title":"BoardCreate","description":"BRD-01: без названия доска получает DEFAULT_TITLE."},"BoardMove":{"properties":{"folder_id":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"Folder Id"}},"additionalProperties":false,"type":"object","required":["folder_id"],"title":"BoardMove","description":"BRD-10: перенос доски в папку; null — на верхний уровень."},"BoardOut":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"title":{"type":"string","title":"Title"},"folder_id":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"Folder Id"},"favorite":{"type":"boolean","title":"Favorite"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","required":["id","title","folder_id","favorite","created_at","updated_at"],"title":"BoardOut"},"BoardRename":{"properties":{"title":{"type":"string","maxLength":200,"minLength":1,"title":"Title"}},"additionalProperties":false,"type":"object","required":["title"],"title":"BoardRename","description":"BRD-02"},"BoardSort":{"type":"string","enum":["updated","created","title"],"title":"BoardSort","description":"BRD-05: порядок полного списка. Даты — от новых к старым, название — по алфавиту."},"Credentials":{"properties":{"email":{"type":"string","maxLength":254,"title":"Email"},"password":{"type":"string","maxLength":1024,"title":"Password"}},"type":"object","required":["email","password"],"title":"Credentials","description":"Почта и пароль для входа; формат почты не проверяется, чтобы отказ был одинаковым."},"FolderCreate":{"properties":{"title":{"type":"string","maxLength":200,"minLength":1,"title":"Title"},"parent_id":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"Parent Id"}},"additionalProperties":false,"type":"object","required":["title"],"title":"FolderCreate","description":"BRD-09: новая папка встаёт последней среди соседей."},"FolderMove":{"properties":{"parent_id":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"Parent Id"},"position":{"type":"integer","minimum":0.0,"title":"Position"}},"additionalProperties":false,"type":"object","required":["parent_id","position"],"title":"FolderMove","description":"BRD-10: новый родитель и место среди его папок (0 — первой; больше числа — последней)."},"FolderOut":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"parent_id":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"Parent Id"},"title":{"type":"string","title":"Title"},"position":{"type":"integer","title":"Position"},"favorite":{"type":"boolean","title":"Favorite"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","parent_id","title","position","favorite","created_at"],"title":"FolderOut"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Health":{"properties":{"status":{"type":"string","title":"Status"}},"type":"object","required":["status"],"title":"Health"},"UserCreate":{"properties":{"name":{"type":"string","maxLength":200,"minLength":1,"title":"Name"},"email":{"type":"string","maxLength":254,"title":"Email","examples":["user@example.com"]},"password":{"type":"string","maxLength":1024,"minLength":1,"title":"Password"}},"additionalProperties":false,"type":"object","required":["name","email","password"],"title":"UserCreate"},"UserOut":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"name":{"type":"string","title":"Name"},"email":{"type":"string","title":"Email"},"disabled":{"type":"boolean","title":"Disabled"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","name","email","disabled","created_at"],"title":"UserOut"},"UserUpdate":{"properties":{"name":{"anyOf":[{"type":"string","maxLength":200,"minLength":1},{"type":"null"}],"title":"Name"},"email":{"anyOf":[{"type":"string","maxLength":254,"examples":["user@example.com"]},{"type":"null"}],"title":"Email"},"password":{"anyOf":[{"type":"string","maxLength":1024,"minLength":1},{"type":"null"}],"title":"Password"},"disabled":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Disabled"}},"additionalProperties":false,"type":"object","title":"UserUpdate","description":"Изменение учётки: переданные поля заменяются, остальные не трогаются."},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}